Refactor sample fixture data to store in redux store
